requirements.txt
# HomeLLM - зависимости для тренировки трансформеров дома
# =========================================================

# PyTorch (>=2.0 для Flash Attention)
torch>=2.0.0

# HuggingFace
transformers>=4.40.0
accelerate>=0.28.0
datasets>=2.19.0
safetensors>=0.4.0

# DeepSpeed (для ZeRO optimizer)
deepspeed>=0.14.0
ninja

# Утилиты
tqdm>=4.66.0
numpy>=1.24.0
pandas>=2.0.0
psutil>=5.9.0
pydantic>=1.10,<2.0
pyyaml>=6.0

# Визуальное приложение (Training Studio)
streamlit>=1.32.0
plotly>=5.18.0
graphviz>=0.20.1

# Логирование и мониторинг
wandb>=0.17.0
tensorboard>=2.16.0
